pyproject.toml
[tool.poetry]
name = "dr-doc-search"
version = "0.11.0"
description = "Search through a document using a chat interface"
authors = [
    "namuan <github@deskriders.dev>",
]
license = "MIT"
readme = "README.md"

documentation = "https://namuan.github.io/dr-doc-search"
homepage = "https://namuan.github.io/dr-doc-search"
repository = "https://github.com/namuan/dr-doc-search"

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

packages = [
    { include = "doc_search", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.9.0, <4.0"
py-executable-checklist = "1.3.1"
rich = "^13.0.0"
pypdf = "^3.2.0"
pytest = "^7.2.0"
openai = "^0.25.0"
langchain = "^0.0.57"
faiss-cpu = "^1.7.3"
python-dotenv = "^0.21.0"

[tool.poetry.group.dev.dependencies]
autoflake = "*"
black = "*"
flake8 = "*"
flake8-bugbear = "*"
flake8-builtins = "*"
flake8-comprehensions = "*"
flake8-debugger = "*"
flake8-logging-format = "*"
isort = "*"
mkdocstrings = "*"
mkdocs-material = "9.0.1"
mypy = "*"
pep8-naming = "*"
pre-commit = "*"
pymdown-extensions = "*"
python-kacl = "*"
pyupgrade = "*"
tryceratops = "*"
commitizen = "^2.20.3"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
coverage = "^7.0.3"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.11.0"
tag_format = "$version"
version_files = [
    "pyproject.toml:version",
]
update_changelog_on_bump = true

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]

[tool.black]
line-length = 120
target-version = ["py37", "py38", "py39"]
include = '\.pyi?$'

[tool.pytest.ini_options]
addopts = """\
    --cov doc_search \
    --cov-report term-missing \
    --no-cov-on-fail \
"""

[tool.coverage.report]
fail_under = 70
omit = ["src/doc_search/app.py", "src/doc_search/__init__.py"]
exclude_lines = [
    'pragma: no cover'
]

[tool.mypy]
disallow_any_unimported = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_equality = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
check_untyped_defs = true
show_error_codes = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dr-doc-search = 'doc_search.app:main'
